requirements.txt
Twisted==18.9.0
lxml==4.2.5
incremental==17.5.0
pyOpenSSL=18.0.0
Scrapy==1.5.0
requests==2.20.0
PyExecJS==1.5.1

# nodejs